requirements/app.txt
repobee>=3.0.0-alpha.1
gitpython==3.1.3
